refactor some code with cpplinter
