+++ src/json/json_functions.cpp
