--- src/math_utils/BigNum.cpp
@@ -1,3 +1,0 @@
-#include "math_utils/BigNum.h"
-
-namespace cuhksz {}  // namespace cuhksz
