--- src/math_utils/Fraction.cpp
@@ -1,3 +1,0 @@
-#include "math_utils/Fraction.h"
-
-namespace cuhksz {}  // namespace cuhksz
